% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-matrix.R
\name{dcm_cat}
\alias{dcm_cat}
\title{Concatenate a list of matrix blocks}
\usage{
dcm_cat(x, d = NULL)
}
\arguments{
\item{x}{A list (or matrix of lists) of numeric blocks.}

\item{d}{Optional dimension along which to concatenate.}
}
\value{
A single matrix.
}
\description{
Concatenate a list of matrix blocks
}
\keyword{internal}
