% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-matrix.R
\name{dcm_diag}
\alias{dcm_diag}
\title{Block-diagonal layout from a list}
\usage{
dcm_diag(...)
}
\arguments{
\item{...}{A list and an optional offset \code{K}.}
}
\value{
A list (or matrix) with the inputs on the diagonal.
}
\description{
Block-diagonal layout from a list
}
\keyword{internal}
