% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{dcm_fieldindices}
\alias{dcm_fieldindices}
\title{Locate or describe entries in a structured parameter}
\usage{
dcm_fieldindices(X, ...)
}
\arguments{
\item{X}{A structured parameter (list with numeric fields).}

\item{...}{One or more field names or numeric indices.}
}
\value{
Integer indices or a character description.
}
\description{
Given a structured parameter object \code{X} (with named fields), returns
the linear indices of one or more named fields, or the structural location
of a given linear index.
}
\keyword{internal}
