% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{dcm_find_pC}
\alias{dcm_find_pC}
\title{Find indices of free parameters in a DCM}
\usage{
dcm_find_pC(...)
}
\arguments{
\item{...}{Either \code{(DCM, fields)} or \code{(pC, pE, fields)}.}
}
\value{
List with \code{i} (free indices), plus \code{pC}, \code{pE}, \code{Np}.
}
\description{
Find indices of free parameters in a DCM
}
\keyword{internal}
