% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{dcm_find_rC}
\alias{dcm_find_rC}
\title{Recover priors from a DCM container}
\usage{
dcm_find_rC(DCM)
}
\arguments{
\item{DCM}{A DCM list.}
}
\value{
List with \code{pC}, \code{pE}.
}
\description{
Recover priors from a DCM container
}
\keyword{internal}
