% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expm.R
\name{expv}
\alias{expv}
\title{Krylov-based action of matrix exponential}
\usage{
expv(t, A, v, tol = 1e-07, m = NULL)
}
\arguments{
\item{t}{Step length.}

\item{A}{Square matrix.}

\item{v}{Vector.}

\item{tol}{Tolerance.}

\item{m}{Krylov subspace dimension.}
}
\value{
Vector \code{expm(t*A) \%*\% v}.
}
\description{
Computes \code{w = expm(t*A) \%*\% v} without forming \code{expm(t*A)}.
Used as a fallback in \code{\link{dcm_dx}} for very large systems.
}
\keyword{internal}
