% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expm.R
\name{padm}
\alias{padm}
\title{Pade approximation of the matrix exponential}
\usage{
padm(A, p = 6)
}
\arguments{
\item{A}{Square matrix.}

\item{p}{Order.}
}
\value{
Matrix exponential of \code{A}.
}
\description{
Pure-R Pade approximation, used as a fallback. See also
\code{\link[expm]{expm}}.
}
\keyword{internal}
